+++ pyobjc-core/Modules/objc/class-builder.m
@@ -356,7 +356,7 @@ exit:
             }
             tc2tc(buf);
             // XXX: The line below doesn't do anything...
-            buf = (char*)PyObjCRT_SkipTypeSpec(buf);
+            // buf = (char*)PyObjCRT_SkipTypeSpec(buf);
         }
         break;